test detection now traverses directory tree. test detetion methods exposed on module level. -U renamed -a. default_tests configuration added.
