+++ booktest/cli.py
@@ -4,57 +4,18 @@ It can be used for creating insights or for creating topics.
 """
 
 import argparse
-import os.path
 import argcomplete
 
-import os
-import importlib
 import sys
-from inspect import signature, Parameter
 
 import booktest as bt
 from booktest.config import get_default_config
+from booktest.detection import detect_tests
 
 
 def add_exec(parser, method):
     parser.set_defaults(
         exec=method)
-
-
-def detect_tests(path, include_in_sys_path=False):
-    tests = []
-    if os.path.exists(path):
-        if include_in_sys_path:
-            sys.path.insert(0, os.path.curdir)
-
-        for f in os.listdir(path):
-            if f.endswith("_test.py") or f.endswith("_book.py") or f.endswith("_suite.py") or \
-                    (f.startswith("test_") and f.endswith(".py")):
-                module_name = os.path.join(path, f[:len(f) - 3]).replace("/", ".")
-                module = importlib.import_module(module_name)
-                for name in dir(module):
-                    member = getattr(module, name)
-                    if isinstance(member, type) and \
-                            issubclass(member, bt.TestBook):
-                        member_signature = signature(member)
-                        needed_arguments = 0
-                        for parameter in member_signature.parameters.values():
-                            if parameter.default == Parameter.empty:
-                                needed_arguments += 1
-                        if needed_arguments == 0:
-                            tests.append(member())
-                    elif isinstance(member, bt.TestBook) or \
-                            isinstance(member, bt.Tests):
-                        tests.append(member)
-
-    return tests
-
-
-def detect_test_suite(path, include_in_sys_path=False):
-    tests = detect_tests(path, include_in_sys_path)
-
-    return bt.merge_tests(tests)
-
 
 def setup_test_suite(parser):
     config = get_default_config()
