+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
 name = "booktest"
-version = "0.3.2"
+version = "0.3.3"
 authors = ["Antt Rauhala <antti@lumoa.io>"]
 description = "Booktest is a snapshot testing library for review driven testing."
 readme = "readme.md"
